pyproject.toml
[tool.poetry]
name = "make-thread-summary"
version = "0.1.0"
description = ""
authors = ["Yuki Fujita <uni22yu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
